requirements-azure.txt
# Azure Functions Requirements - Minimal and Reliable
azure-functions>=1.18.0
azure-functions-worker>=1.0.15

# Core API
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Data Processing  
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# Financial Data
yfinance>=0.2.0

# ML (basic only)
scikit-learn>=1.3.0
joblib>=1.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0